Move TreeBinarySerializer to the correct namespace
