+++ SAGESharp.Tests/SAGESharp.Tests.csproj
@@ -85,6 +85,7 @@
     <Compile Include="IO\Binary\BinarySerializerTests.cs" />
     <Compile Include="IO\Binary\BinaryWriterWrapperTests.cs" />
     <Compile Include="IO\Binary\ReaderTests.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\Binary\WriterTests.cs" />
     <Compile Include="IO\EdgeTests.cs" />
     <Compile Include="IO\ListNodeTests.cs" />
@@ -93,7 +94,6 @@
     <Compile Include="IO\PrimitiveTypeDataNodeTests.cs" />
     <Compile Include="IO\StringDataNodeTests.cs" />
     <Compile Include="IO\Substitutes.cs" />
-    <Compile Include="IO\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\TreeBuilderTests.cs" />
     <Compile Include="IO\TreeReaderTests.cs" />
     <Compile Include="IO\Trees\Builders.cs" />
